src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anndata>=0.11.4",
    "cython>=3.0.12",
    "jupyterlab>=4.3.6",
    "legacy-api-wrap>=1.4.1",
    "notebook>=7.3.3",
    "numba>=0.61.0",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "pyqt5>=5.15.11",
    "ripser>=0.6.12",
    "scanpy>=1.11.1",
    "scikit-learn>=1.5.2",
    "scikit-network>=0.33.1",
    "seaborn>=0.13.2",
    "tables>=3.10.1",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.7",
]
